requirements.txt
streamlit
streamlit_mic_recorder
sounddevice
numpy
scipy
google-generativeai
python-dotenv
gtts
pygame
